+++ Type/Boolean.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Tdn\PhpTypes\Type;
+
+/**
+ * Class Boolean.
+ *
+ * @deprecated
+ */
+class Boolean extends BooleanType
+{
+}