+++ Type/String.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Tdn\PhpTypes\Type;
+
+/**
+ * Class String.
+ *
+ * @deprecated
+ */
+class String extends StringType
+{
+}